fix(plugin-page-builder): check a stored tag before handing it to the renderer
Validation asks only that props be an object, so `as` reaches render as data.
A stored void element handed children throws inside React and takes the page
with it, so an unrecognised tag now falls back to a div through a real type
guard rather than being trusted because the type said so.

The exhaustive support-key check sat in a `*.test.ts`, which the package's
tsconfig excludes, so the half of that guard I described as compile-time never
ran: Vitest only transpiles. It moves to a `*.test-d.ts`, which the project
does include, and the augmentation proof moves to plugin-page-builder, which
consumes the SDK the way a plugin does. Keeping the augmentation beside the
guard would have widened the very type the guard measures.

The contained-container class has no rule behind it anywhere yet, so a section
promising the site's content width rendered full width. The class is the seam
and the rule belongs to the site stylesheet where its token will live; both the
code and the block's own description now say that instead of promising width.

The changeset now names what this removes, since the revision helpers were
published: they duplicated the content versioning already in core.
